+++ docs/changelog.md
@@ -2,6 +2,8 @@
 ======================================================
 **API 2.0.0-SNAPSHOT**
 **Requires Journeymap v6.0.0**
+* Added Paper support: `journeymap-api-paper` artifact for writing server-side JourneyMap addons as Paper plugins.
+* Added server plugin support and server api and events.
 * Added Server Properties, so mods can add properties to JourneyMap's Server Admin options tab.
 * Added LayeredScreen abstraction for layirng screens.
 * Refactored Waypoint and WaypointGroups to be more concrete.